API

Modules

BATTestCases.BATTestCases

Types and constants

BATTestCases.FunnelDistribution

BATTestCases.GaussianShell

BATTestCases.MultimodalStudentT

Functions and macros

Documentation

BATTestCases.BATTestCases — Module

BATTestCases

A collection of test cases for BAT.jl.

BATTestCases.FunnelDistribution — Type

struct FunnelDistribution <: Distribution{Multivariate,Continuous}

Experimental feature, not part of stable public API.

A funnel distribution (see Caldwell et al. for definition).

Constructors:

FunnelDistribution(; a::Real = 1.0, b::Real = 0.5, n::Integer = 3)

Fields:

a::Real: Variance of the dominant normal distribution.

b::Real: Variance of the supporting normal distributions.

n::Integer: Number of dimensions.

BATTestCases.GaussianShell — Type

struct GaussianShell <: Distribution{Multivariate,Continuous}

Experimental feature, not part of stable public API.

Gaussian Shell (see Caldwell et al. for definition).

Constructors:

GaussianShell(; r::Real=5, w::Real=2, n::Integer=2)

Fields:

r::Real: The radius of the Gaussian shell distribution.

w::Real: Variance of the Gaussian shell distribution

n::Int64: Number of dimensions

c::AbstractVector{<:Real}

lognorm::AbstractFloat

Note

Fields c and lognorm are considered internal and subject to change without deprecation.

BATTestCases.MultimodalStudentT — Type

MultimodalStudentT <: Distribution{Multivariate,Continuous}

The Multimodal Student-t Distribution (It's defined similarly to multimodal Cauchy defined in Caldwell et al.).

Assumes two bimodal peaks, each in its own dimension.

Constructors:

MultimodalStudentT(; μ::Real=1, σ::Float64=0.2, ν::Integer=1, n::Integer=4)

Arguments:

μ::Real: The location parameter used for the two bimodal peaks.

σ::Float64: The scale parameter shared among all components.

ν::Int: The degrees of freedom.

n::Int: The number of dimensions.

Fields:

bimodals::Distributions.MixtureModel

σ::Float64

n::Int64

dist::Distributions.Product

Note

All fields of MultimodalStudentT are considered internal and subject to change without deprecation.